pyproject.toml
[tool.poetry]
name = "everythird"
version = "0.1.0"
description = ""
authors = ["jassycliq <josalgado12709@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^4.0.4"
djangorestframework = "^3.13.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
